Fix some instances of -Wsign-compare and -Wignored-qualifiers
Fix instances in chrome/browser/ash/ as a precursor to reenable those
warnings in //third_party/shell-encryption.

Bug: 1095122
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3921579
Commit-Queue: Henrique Ferreiro <hferreiro@igalia.com>
Cr-Commit-Position: refs/heads/main@{#1052317}

+++ chrome/browser/ash/arc/input_overlay/actions/action.cc
@@ -5,6 +5,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 
 #include "chrome/browser/ash/arc/input_overlay/actions/action.h"
 
+#include "base/check_op.h"
 #include "chrome/browser/ash/arc/input_overlay/actions/position.h"
 #include "chrome/browser/ash/arc/input_overlay/touch_id_manager.h"
 #include "ui/aura/window.h"
@@ -394,11 +395,11 @@ bool Action::VerifyOnKeyRelease(ui::DomCode code) {
   if (!touch_id_) {
     LOG(ERROR) << "There should be a touch ID for the release {"
                << ui::KeycodeConverter::DomCodeToCodeString(code) << "}.";
-    DCHECK_EQ(keys_pressed_.size(), 0);
+    DCHECK_EQ(keys_pressed_.size(), 0u);
     return false;
   }
 
-  DCHECK_NE(keys_pressed_.size(), 0);
+  DCHECK_NE(keys_pressed_.size(), 0u);
   if (keys_pressed_.size() == 0 || !keys_pressed_.contains(code))
     return false;
 
@@ -462,7 +463,7 @@ void Action::UpdateTouchDownPositions() {
 
   touch_down_positions_.clear();
   const auto& content_bounds = touch_injector_->content_bounds();
-  for (int i = 0; i < original_positions_.size(); i++) {
+  for (size_t i = 0; i < original_positions_.size(); i++) {
     auto point = beta_
                      ? current_positions_[i].CalculatePosition(content_bounds)
                      : original_positions_[i].CalculatePosition(content_bounds);
